feat: enable filter touch on android
We had a customer asking for the addition of a configuration
to our Android’s manifest to prevent touches when the app is
obscured, as this was flagged by the google store as a
security issue.

Therefore, we should provide the ability for our customers to
add this specific configuration.

References https://outsystemsrd.atlassian.net/browse/RNMT-5303