feat: re-introduce legacy Cordova splash screen logic
The changes made to splash screens in Cordova-android 11
are in line with the UX Google wants. Unfortunately, this removes
the support for custom full image splash screens altogether.

To keep the support for the new Splash Screen API provided
out of the box by Cordova but also support the full image splash
screens, we're reintroducing the removed code while keeping
the new logic as well.

So what does this do now?
- If no custom splash screens were provided and the
ShowSplashScreenSpinner preference is not enabled, only the
new Android splash is presented
- Otherwise, the Android splash is dismissed as ASAP and the
old experience is used
- The show() method is now available in the clobber once again,
because the old splash screen can still be re-shown (the new one
can't, hence why this method was removed)
- It's possible to dismiss the splash screen via hide() when using
AutoHideSplashScreen (this was possible before but was removed
in this new experience)

References https://outsystemsrd.atlassian.net/browse/RNMT-5730
References https://outsystemsrd.atlassian.net/browse/RNMT-6187