feat: ensure Kotlin JVM target is 1.8
As per https://developer.android.com/build/jdks#target-compat, when
using Kotlin, the jvmTarget defaults to the toolchain version which,
if not explicit, is the JDK used by Gradle. Now that AGP 8 requires
JDK 17 (which is also embedded in the latest Android Studio version
[Giraffe as of this commit]), the default toolchain is effectively
JDK 17 and this is causing JVM target incompatibility when building
an application using both Java and Kotlin source code. The error is
something along the lines of:

 'compileDebugJavaWithJavac' task (current target is 1.8) and
 'kaptGenerateStubsDebugKotlin' task (current target is 17)
 jvm target compatibility should be set to the same Java version.

As per https://kotlinlang.org/docs/gradle-configure-project.html#what-can-go-wrong-if-not-checking-targets-compatibility
this incompatibility can arise when:

 - Have a default configuration, and your JDK is not equal to 1.8.

Since the build.gradle file is setting the Java source and target
compatibility to 1.8, this causes a mismatch with the new default
assumed by Kotlin's jvmTarget. There are two ways out of this:

 1. Raise the Java source and target compatibility values to 17 and
    do nothing about Kotlin's jvmTarget
 2. Explictly set Kotlin's jvmTarget to 1.8 and do nothing about
    Java source and target compatibility values

Point 2. was chosen because although supposedly Android 14 will
bring full support for Java 17 language features according to
https://developer.android.com/about/versions/14/summary, currently
any fresh project generated via Android Studio still sets version
1.8 for Java source and target compatibility, and actually lowers
Kotlin's jvmTarget to 1.8, meaning Google is not ready yet to fully
pull the trigger on Java 17, and neither should we.

References https://outsystemsrd.atlassian.net/browse/RNMT-6064